# schemas/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisRequest": {
        "additionalProperties": false,
        "properties": {
          "inventory_id": {
            "title": "Inventory Id",
            "type": "string"
          },
          "leaf_id": {
            "title": "Leaf Id",
            "type": "string"
          },
          "target_context": {
            "$ref": "#/components/schemas/TargetContext"
          }
        },
        "required": [
          "inventory_id",
          "leaf_id",
          "target_context"
        ],
        "title": "AnalysisRequest",
        "type": "object"
      },
      "AnalysisResponse": {
        "additionalProperties": false,
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/AnalysisResult"
          }
        },
        "required": [
          "analysis"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
      "AnalysisResult": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "evidence": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EvidenceSpan"
                },
                {
                  "$ref": "#/components/schemas/DossierEvidence"
                }
              ]
            },
            "title": "Evidence",
            "type": "array"
          },
          "expert_validated": {
            "default": false,
            "title": "Expert Validated",
            "type": "boolean"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "human_approval_required": {
            "title": "Human Approval Required",
            "type": "boolean"
          },
          "id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Id",
            "type": "string"
          },
          "model_run": {
            "$ref": "#/components/schemas/ModelRunRecord",
            "default": {
              "latency_ms": 0.0,
              "mode": "disabled",
              "model_name": "contract-default",
              "prompt_template_version": "1.0.0",
              "status": "not_applicable"
            }
          },
          "operational_status": {
            "$ref": "#/components/schemas/OperationalStatus"
          },
          "rationale": {
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "repair": {
            "$ref": "#/components/schemas/RepairAction"
          },
          "scenario_disclosure": {
            "minLength": 1,
            "title": "Scenario Disclosure",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "source_artifact": {
            "$ref": "#/components/schemas/SourceArtifact"
          },
          "target_context": {
            "$ref": "#/components/schemas/TargetContext"
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/TraceStep"
            },
            "minItems": 1,
            "title": "Trace",
            "type": "array"
          },
          "triggered_rule_ids": {
            "items": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
              "type": "string"
            },
            "title": "Triggered Rule Ids",
            "type": "array"
          },
          "unresolved_uncertainty": {
            "items": {
              "type": "string"
            },
            "title": "Unresolved Uncertainty",
            "type": "array"
          }
        },
        "required": [
          "id",
          "source_artifact",
          "target_context",
          "operational_status",
          "scenario_disclosure",
          "decision",
          "severity",
          "triggered_rule_ids",
          "findings",
          "evidence",
          "rationale",
          "repair",
          "confidence",
          "unresolved_uncertainty",
          "human_approval_required",
          "trace"
        ],
        "title": "AnalysisResult",
        "type": "object"
      },
      "ApplicationInventory": {
        "additionalProperties": false,
        "properties": {
          "applicant_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicant Name"
          },
          "application_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Number"
          },
          "fixture_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixture Id"
          },
          "has_stf": {
            "title": "Has Stf",
            "type": "boolean"
          },
          "id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Id",
            "type": "string"
          },
          "leaves": {
            "items": {
              "$ref": "#/components/schemas/ParsedLeaf"
            },
            "minItems": 1,
            "title": "Leaves",
            "type": "array"
          },
          "package_sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Package Sha256",
            "type": "string"
          },
          "source_standard": {
            "$ref": "#/components/schemas/StandardVersion"
          },
          "submission_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Type"
          },
          "warnings": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ParseWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "source_standard",
          "has_stf",
          "package_sha256",
          "leaves"
        ],
        "title": "ApplicationInventory",
        "type": "object"
      },
      "ApplicationType": {
        "enum": [
          "NDA",
          "ANDA",
          "IND",
          "DMF"
        ],
        "title": "ApplicationType",
        "type": "string"
      },
      "Authority": {
        "enum": [
          "FDA"
        ],
        "title": "Authority",
        "type": "string"
      },
      "BaselineRunRequest": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "system": {
            "enum": [
              "B0",
              "B1",
              "B2",
              "RegBridge"
            ],
            "title": "System",
            "type": "string"
          }
        },
        "required": [
          "system",
          "case_id"
        ],
        "title": "BaselineRunRequest",
        "type": "object"
      },
      "BaselineRunResponse": {
        "additionalProperties": false,
        "properties": {
          "current_fda_operational_availability": {
            "const": "not_operational",
            "title": "Current Fda Operational Availability",
            "type": "string"
          },
          "eligible_for_performance_claims": {
            "const": false,
            "title": "Eligible For Performance Claims",
            "type": "boolean"
          },
          "empirical_model_run": {
            "const": false,
            "title": "Empirical Model Run",
            "type": "boolean"
          },
          "prediction": {
            "$ref": "#/components/schemas/SystemPrediction"
          },
          "retrieval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalTrace"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_type": {
            "const": "deterministic_fixture_validation",
            "title": "Run Type",
            "type": "string"
          }
        },
        "required": [
          "run_type",
          "empirical_model_run",
          "eligible_for_performance_claims",
          "current_fda_operational_availability",
          "prediction"
        ],
        "title": "BaselineRunResponse",
        "type": "object"
      },
...
      "Center": {
        "enum": [
          "CDER"
        ],
        "title": "Center",
        "type": "string"
      },
...
      "EvaluationCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "configuration_id": {
            "const": "m3-fixture-all-systems-v1",
            "title": "Configuration Id",
            "type": "string"
          }
        },
        "required": [
          "configuration_id"
        ],
        "title": "EvaluationCreateRequest",
        "type": "object"
      },
      "EvaluationResponse": {
        "additionalProperties": false,
        "properties": {
          "evaluation": {
            "$ref": "#/components/schemas/EvaluationRun"
          }
        },
        "required": [
          "evaluation"
        ],
        "title": "EvaluationResponse",
        "type": "object"
      },
...
      "GraphNeighborhood": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Analysis Id",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "minItems": 1,
            "title": "Nodes",
            "type": "array"
          },
          "text_alternative": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Text Alternative",
            "type": "array"
          }
        },
        "required": [
          "analysis_id",
          "nodes",
          "edges",
          "text_alternative"
        ],
        "title": "GraphNeighborhood",
        "type": "object"
      },
...
      "GraphResponse": {
        "additionalProperties": false,
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/GraphNeighborhood"
          }
        },
        "required": [
          "graph"
        ],
        "title": "GraphResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ParsedLeaf": {
        "additionalProperties": false,
        "properties": {
          "content_type": {
            "pattern": "^[a-z0-9][a-z0-9.+-]*/[a-z0-9][a-z0-9.+-]*$",
            "title": "Content Type",
            "type": "string"
          },
          "declared_checksum": {
            "anyOf": [
              {
                "pattern": "^[a-f0-9]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Declared Checksum"
          },
          "extraction_status": {
            "default": "completed",
            "enum": [
              "completed",
              "failed",
              "bounded"
            ],
            "title": "Extraction Status",
            "type": "string"
          },
          "file_sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "File Sha256",
            "type": "string"
          },
          "heading": {
            "pattern": "^\\d+(?:\\.[A-Za-z0-9]+)+$",
            "title": "Heading",
            "type": "string"
          },
          "href": {
            "minLength": 1,
            "title": "Href",
            "type": "string"
          },
          "hyperlink_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Hyperlink Count",
            "type": "integer"
          },
          "id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Id",
            "type": "string"
          },
          "keywords": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ParsedKeyword"
            },
            "title": "Keywords",
            "type": "array"
          },
          "modified_leaf_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified Leaf Id"
          },
          "operation": {
            "$ref": "#/components/schemas/LifecycleOperation"
          },
          "source_locator": {
            "minLength": 1,
            "title": "Source Locator",
            "type": "string"
          },
          "text_span_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Text Span Count",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "heading",
          "href",
          "operation",
          "content_type",
          "file_sha256",
          "source_locator"
        ],
        "title": "ParsedLeaf",
        "type": "object"
      },
...
      "StandardVersion": {
        "enum": [
          "eCTD-3.2.2",
          "eCTD-4.0"
        ],
        "title": "StandardVersion",
        "type": "string"
      },
...
      "TargetContext": {
        "additionalProperties": false,
        "properties": {
          "analysis_date": {
            "format": "date",
            "title": "Analysis Date",
            "type": "string"
          },
          "application_type": {
            "$ref": "#/components/schemas/ApplicationType"
          },
          "authority": {
            "$ref": "#/components/schemas/Authority"
          },
          "center": {
            "$ref": "#/components/schemas/Center"
          },
          "metadata_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetadataPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "reuse_operation": {
            "$ref": "#/components/schemas/ReuseOperation"
          },
          "scenario_mode": {
            "$ref": "#/components/schemas/ScenarioMode"
          },
          "source_standard": {
            "$ref": "#/components/schemas/StandardVersion"
          },
          "standards_snapshot_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Standards Snapshot Id",
            "type": "string"
          },
          "target_standard": {
            "$ref": "#/components/schemas/StandardVersion"
          }
        },
        "required": [
          "authority",
          "center",
          "application_type",
          "source_standard",
          "target_standard",
          "analysis_date",
          "reuse_operation",
          "standards_snapshot_id",
          "scenario_mode"
        ],
        "title": "TargetContext",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Decision-support contracts for evidence-grounded, version-aware analysis. RegBridge is not FDA-certified and does not predict submission acceptance.",
    "summary": "FDA/CDER-scoped legacy eCTD reuse risk-analyzer research prototype.",
    "title": "RegBridge API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyses": {
      "post": {
        "operationId": "create_analysis_api_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Analysis",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "operationId": "get_analysis_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analysis",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/analyses/{analysis_id}/graph": {
      "get": {
        "operationId": "get_analysis_graph_api_v1_analyses__analysis_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analysis Graph",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/applications/parse": {
      "post": {
        "operationId": "parse_application_api_v1_applications_parse_post",
        "parameters": [
          {
            "in": "query",
            "name": "fixture_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fixture Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationInventory"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parse Application",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/baselines/run": {
      "post": {
        "operationId": "run_baseline_api_v1_baselines_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Baseline",
        "tags": [
          "evaluation"
        ]
      }
    },
...
    "/api/v1/evaluations": {
      "post": {
        "operationId": "create_evaluation_api_v1_evaluations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Evaluation",
        "tags": [
          "evaluation"
        ]
      }
    },
    "/api/v1/evaluations/{evaluation_id}": {
      "get": {
        "operationId": "get_evaluation_api_v1_evaluations__evaluation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_id",
            "required": true,
            "schema": {
              "title": "Evaluation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evaluation",
        "tags": [
          "evaluation"
        ]
      }
    },
...
  }
}